chore: move plugin into router namespace
