+++ cli/src/commands/router/plugin/index.ts
@@ -1,5 +1,5 @@
 import { Command } from 'commander';
-import { BaseCommandOptions } from '../../core/types/types.js';
+import { BaseCommandOptions } from '../../../core/types/types.js';
 import InitPluginCommand from './commands/init.js';
 import BuildPluginCommand from './commands/build.js';
 import TestPluginCommand from './commands/test.js';
