+++ cli/src/commands/router/plugin/templates/go-plugin.ts
